pyproject.toml
[project]
name = "supernote-stickers"
version = "0.1.0"
description = "Convert images to Supernote SNSTK sticker pack format"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Supernote Fan", email = "supernote-stickers@users.noreply.github.com" }]
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
    "flask>=3.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
png2snstk = "supernote_stickers.cli:main"
snstk-web = "supernote_stickers.web.app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/supernote_stickers"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
